pyproject.toml
[project]
name = "like-sticker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "nicegui>=3.11.1",
    "pillow>=12.2.0",
    "rembg[cpu]>=2.0.75",
]
